token tracking and ui stuff
