+++ components/MessageCard.tsx
@@ -183,7 +183,7 @@ const MessageCard: React.FC<MessageCardProps> = ({
     if (isRegenerateAction && !aboutToRegenerate) {
       aboutToRegenerate = true;
       vibrate(50);
-      console.log("Message card - regen vibrate...")
+      // console.log("Message card - regen vibrate...")
       //debugger;
     } else if (!isRegenerateAction) {
       aboutToRegenerate = false;
@@ -195,13 +195,13 @@ const MessageCard: React.FC<MessageCardProps> = ({
         : 0,
       y: 0,
       height: 100,
-      blur: isRegenerateAction ? 2 : 0,
+      blur: isRegenerateAction ? 4 : (down ? 0.01 * -mx : 0),
       fontSize: 1,
       config: { tension: 190, friction: 18 },
     });
 
     apiScaleSpring.start({
-      scale: down ? (isRegenerateAction ? 0.8 : 0.95) : 1,
+      scale: down ? (isRegenerateAction ? 0.95 : 1) : 1,
     });
   
     if (((vx > 2 && mx < 0) || mx < dragThreshold) && !down && isEligibleForRegenerate) {
@@ -551,16 +551,16 @@ const MessageCard: React.FC<MessageCardProps> = ({
                 }}>
                   {renderContent()}
                   {/* Swipe to regenerate overlay */}
-                  <animated.div className="absolute inset-0 bg-black flex gap-2 items-end justify-end z-10 text-white bg-opacity-0 pointer-events-none select-none"
-                   style={{
-                     opacity: canRegenerate ? x.to(val => -val / 100) : 0
-                   }}>
-                   
-                   <RotateCw className="animate-[spin_2s_infinite]" />
-                   <p>Swipe left to rewrite...</p>
-                  </animated.div>
                 </animated.div>
               </CardContent>
+              <animated.div className="absolute inset-0 bg-black flex gap-2 items-end justify-end z-10 text-white bg-opacity-0 pointer-events-none select-none"
+                style={{
+                  opacity: canRegenerate ? x.to(val => -val / 100) : x.to(() => 0)
+                }}>
+                
+                <RotateCw className="animate-[spin_2s_infinite]" />
+                <p>Swipe left to rewrite...</p>
+              </animated.div>
             </Card>
             {/* <AnimatePresence>
             {!globalIsThinking && isLastMessage && role === "assistant" && !isEditing && (
